+++ src/main/java/com/leverx/javacourse/seller/rating/app/SellerRatingApplication.java
@@ -1,0 +1,13 @@
+package com.leverx.javacourse.seller.rating.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SellerRatingApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SellerRatingApplication.class, args);
+	}
+
+}
