+++ src/main/java/com/leverx/javacourse/seller/rating/app/controller/UserController.java
@@ -1,0 +1,100 @@
+package com.leverx.javacourse.seller.rating.app.controller;
+
+import com.leverx.javacourse.seller.rating.app.dto.UserCreateDto;
+import com.leverx.javacourse.seller.rating.app.dto.UserResponseDto;
+import com.leverx.javacourse.seller.rating.app.mapper.UserDtoMapper;
+import com.leverx.javacourse.seller.rating.app.entity.model.Comment;
+import com.leverx.javacourse.seller.rating.app.entity.model.Seller;
+import com.leverx.javacourse.seller.rating.app.entity.model.User;
+import com.leverx.javacourse.seller.rating.app.entity.model.UserRoles;
+import com.leverx.javacourse.seller.rating.app.service.CommentService;
+import com.leverx.javacourse.seller.rating.app.service.UserService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.math.BigDecimal;
+import java.util.List;
+
+@RestController
+public class UserController {
+
+    private final UserService userService;
+    private final CommentService commentService;
+    private final UserDtoMapper userDtoMapper;
+
+    public UserController(UserService userService, CommentService commentService, UserDtoMapper userDtoMapper) {
+        this.userService = userService;
+        this.commentService = commentService;
+        this.userDtoMapper = userDtoMapper;
+    }
+
+    @GetMapping("/users/{id}")
+    public ResponseEntity<UserResponseDto> getUser(@PathVariable Long id) {
+        User requestedUser = userService.findById(id);
+        UserResponseDto userResponseDto = userDtoMapper.toUserResponseDto(requestedUser);
+        return ResponseEntity.status(HttpStatus.FOUND).body(userResponseDto);
+    }
+
+    @PostMapping("/auth/register")
+    public ResponseEntity<UserResponseDto> createUser(@RequestBody UserCreateDto userCreateDto) {
+        User user = null;
+        if (UserRoles.SELLER.toString().equals(userCreateDto.getRole())) {
+            user = userService.saveSeller(userDtoMapper.toSeller(userCreateDto));
+        } else if (UserRoles.VISITOR.toString().equals(userCreateDto.getRole())) {
+            user = userService.saveVisitor(userDtoMapper.toVisitor(userCreateDto));
+        }
+        UserResponseDto responseDto = userDtoMapper.toUserResponseDto(user);
+        return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
+    }
+
+    @PostMapping("/auth/register_anon")
+    public ResponseEntity<UserResponseDto> createAnonymousSeller(@RequestBody UserCreateDto userCreateDto, @AuthenticationPrincipal UserDetails userDetails) {
+        Seller seller = userService.saveSeller(userDtoMapper.toSeller(userCreateDto));
+        Comment comment = seller.getAssignedComments().getFirst();
+        comment.setSeller(seller);
+        comment.setAuthor(userService.findByLogin(userDetails.getUsername()).get());
+        commentService.save(comment);
+        UserResponseDto responseDto = userDtoMapper.toUserResponseDto(userService.findById(seller.getId()));
+        return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
+    }
+
+    @GetMapping("/users/all")
+    public ResponseEntity<List<UserResponseDto>> getAllSellersRanked() {
+        List<Seller> allUsers = userService.getSellerByRating();
+        return ResponseEntity.status(HttpStatus.FOUND).body(userDtoMapper.sellerToUserResponseDtoList(allUsers));
+    }
+
+    @GetMapping("/users/all/by_game")
+    public ResponseEntity<List<UserResponseDto>> getUsersByGame(@RequestParam String gameTitle) {
+        List<Seller> usersByGame = userService.getSellersByGame(gameTitle);
+        return ResponseEntity.status(HttpStatus.FOUND).body(userDtoMapper.sellerToUserResponseDtoList(usersByGame));
+    }
+
+    @GetMapping("/users/all/in_range")
+    public ResponseEntity<List<UserResponseDto>> getUsersInRatingRange(@RequestParam BigDecimal begin, @RequestParam BigDecimal end) {
+        List<Seller> requestedUsers = userService.getSellersInRatingRange(begin, end);
+        return ResponseEntity.status(HttpStatus.FOUND).body(userDtoMapper.sellerToUserResponseDtoList(requestedUsers));
+    }
+
+    @DeleteMapping("/users/{id}")
+    public ResponseEntity<String> deleteUser(@PathVariable Long id, @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails.getUsername().equals(userService.findById(id).getLogin()) || userDetails.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .anyMatch(a -> UserRoles.ADMINISTRATOR.getAuthority().equals(a))){
+            userService.deleteById(id);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+    }
+}
