+++ src/main/java/com/leverx/javacourse/seller/rating/app/entity/model/UserRoles.java
@@ -1,0 +1,12 @@
+package com.leverx.javacourse.seller.rating.app.entity.model;
+
+import org.springframework.security.core.GrantedAuthority;
+
+public enum UserRoles implements GrantedAuthority {
+    SELLER, VISITOR, ADMINISTRATOR;
+
+    @Override
+    public String getAuthority() {
+        return name();
+    }
+}
