+++ src/main/java/com/leverx/javacourse/seller/rating/app/exception/EntityNotFoundException.java
@@ -1,0 +1,11 @@
+package com.leverx.javacourse.seller.rating.app.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException() {
+        super();
+    }
+
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
