+++ src/main/java/com/leverx/javacourse/seller/rating/app/repository/CommentRepository.java
@@ -1,0 +1,8 @@
+package com.leverx.javacourse.seller.rating.app.repository;
+
+import com.leverx.javacourse.seller.rating.app.entity.model.Comment;
+import org.springframework.data.repository.CrudRepository;
+
+public interface CommentRepository extends CrudRepository<Comment, Long> {
+
+}
