+++ src/main/java/com/leverx/javacourse/seller/rating/app/repository/ItemRepository.java
@@ -1,0 +1,12 @@
+package com.leverx.javacourse.seller.rating.app.repository;
+
+import com.leverx.javacourse.seller.rating.app.entity.model.Item;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+
+public interface ItemRepository extends CrudRepository<Item, Long> {
+
+    List<Item> findByGameTitle(String gameTitle);
+
+}
