+++ src/main/java/com/leverx/javacourse/seller/rating/app/repository/UserRepository.java
@@ -1,0 +1,21 @@
+package com.leverx.javacourse.seller.rating.app.repository;
+
+import com.leverx.javacourse.seller.rating.app.entity.model.Seller;
+import com.leverx.javacourse.seller.rating.app.entity.model.User;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.CrudRepository;
+
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.Optional;
+
+public interface UserRepository<T extends User> extends CrudRepository<T, Long> {
+
+    @Query("SELECT s FROM Seller s ORDER BY s.rating")
+    List<Seller> findAllSellersByRating();
+
+    @Query("SELECT s FROM Seller s WHERE s.rating BETWEEN :ratingStarts AND :ratingEnds ORDER BY s.rating")
+    List<Seller> findSellerByRatingRange(BigDecimal ratingStarts, BigDecimal ratingEnds);
+
+    Optional<User> findByLogin(String login);
+}
