--- src/main/java/com/leverx/javacourse/seller_rating_app/SellerRatingApplication.java
@@ -1,13 +1,0 @@
-package com.leverx.javacourse.seller_rating_app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SellerRatingApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SellerRatingApplication.class, args);
-	}
-
-}
