--- src/main/java/com/leverx/javacourse/seller_rating_app/entity/dto/ItemCreateDto.java
@@ -1,42 +1,0 @@
-package com.leverx.javacourse.seller_rating_app.entity.dto;
-
-import java.time.LocalDate;
-
-public class ItemCreateDto {
-    private String title;
-    private String description;
-    private String gameTitle;
-    private LocalDate created;
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getGameTitle() {
-        return gameTitle;
-    }
-
-    public void setGameTitle(String gameTitle) {
-        this.gameTitle = gameTitle;
-    }
-
-    public LocalDate getCreated() {
-        return created;
-    }
-
-    public void setCreated(LocalDate created) {
-        this.created = created;
-    }
-}
