--- src/main/java/com/leverx/javacourse/seller_rating_app/entity/dto/ItemResponseDto.java
@@ -1,69 +1,0 @@
-package com.leverx.javacourse.seller_rating_app.entity.dto;
-
-import java.time.LocalDate;
-
-public class ItemResponseDto {
-    private Long Id;
-    private String title;
-    private String description;
-    private String gameTitle;
-    private LocalDate created;
-    private LocalDate updated;
-    private Long seller;
-
-    public Long getId() {
-        return Id;
-    }
-
-    public void setId(Long id) {
-        Id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getGameTitle() {
-        return gameTitle;
-    }
-
-    public void setGameTitle(String gameTitle) {
-        this.gameTitle = gameTitle;
-    }
-
-    public LocalDate getCreated() {
-        return created;
-    }
-
-    public void setCreated(LocalDate created) {
-        this.created = created;
-    }
-
-    public LocalDate getUpdated() {
-        return updated;
-    }
-
-    public void setUpdated(LocalDate updated) {
-        this.updated = updated;
-    }
-
-    public Long getSeller() {
-        return seller;
-    }
-
-    public void setSeller(Long seller) {
-        this.seller = seller;
-    }
-}
