--- src/main/java/com/leverx/javacourse/seller_rating_app/entity/model/UserRoles.java
@@ -1,12 +1,0 @@
-package com.leverx.javacourse.seller_rating_app.entity.model;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum UserRoles implements GrantedAuthority {
-    SELLER, VISITOR, ADMINISTRATOR;
-
-    @Override
-    public String getAuthority() {
-        return name();
-    }
-}
