--- src/main/java/com/leverx/javacourse/seller_rating_app/entity/model/Visitor.java
@@ -1,36 +1,0 @@
-package com.leverx.javacourse.seller_rating_app.entity.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-
-import java.time.LocalDate;
-import java.util.List;
-
-@Entity
-@Table(name = "visitors")
-public class Visitor extends User{
-
-    public Visitor() {
-    }
-
-    public Visitor(Long id, String login, String password, String firstName, String secondName, String email, LocalDate created, List<Comment> ownComments) {
-        super(id, login, password, firstName, secondName, email, created, ownComments, UserRoles.VISITOR);
-    }
-
-    @Override
-    public String toString() {
-        return "Visitor{" +
-                "Id=" + Id +
-                ", comments=" + ownComments +
-                ", Id=" + Id +
-                ", login='" + login + '\'' +
-                ", password=" + password +
-                ", firstName='" + firstName + '\'' +
-                ", secondName='" + secondName + '\'' +
-                ", email='" + email + '\'' +
-                ", created=" + created +
-                '}';
-    }
-}
