--- src/main/java/com/leverx/javacourse/seller_rating_app/exception/EntityNotFoundException.java
@@ -1,11 +1,0 @@
-package com.leverx.javacourse.seller_rating_app.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException() {
-        super();
-    }
-
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
