--- src/main/java/com/leverx/javacourse/seller_rating_app/repository/CommentRepository.java
@@ -1,8 +1,0 @@
-package com.leverx.javacourse.seller_rating_app.repository;
-
-import com.leverx.javacourse.seller_rating_app.entity.model.Comment;
-import org.springframework.data.repository.CrudRepository;
-
-public interface CommentRepository extends CrudRepository<Comment, Long> {
-
-}
