+++ src/main/resources/application.yml
@@ -9,10 +9,6 @@ spring:
   jpa:
     database: postgresql
     defer-datasource-initialization: true
-  security:
-    user:
-      name: Pupa
-      password: Lupa
   sql:
     init:
       mode: always
