+++ src/test/java/com/leverx/javacourse/seller/rating/app/SellerRatingApplicationTests.java
@@ -1,0 +1,13 @@
+package com.leverx.javacourse.seller.rating.app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SellerRatingApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
