--- src/test/java/com/leverx/javacourse/seller_rating_app/SellerRatingApplicationTests.java
@@ -1,13 +1,0 @@
-package com.leverx.javacourse.seller_rating_app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SellerRatingApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
